Skip installation if the dotfile is already installed

+++ install.sh
@@ -4,8 +4,8 @@
 DOT_ROOT=$(cd `dirname "$0"`; pwd)
 
 _confirm() {
+    confirmed="no"
 	local dest_file="$1"
-	confirmed="no"
 
 	if [ -e "$dest_file" -o -L "$dest_file" ]
 	then
@@ -29,11 +29,15 @@ _confirm() {
 _install() {
 	local src_file="$1"
 	local dest_file="$2"
-	_confirm "$dest_file"
-	if [ "$confirmed" == "yes" ]; then
-		[ -e "$dest_file" -o -L "$dest_file" ] && rm "$dest_file"
-		ln -s "$src_file" "$dest_file"
-		echo "Installed $dest_file"
+	if [ "$src_file" -ef "$dest_file" ]; then
+		echo "Skipped $dest_file: already installed"
+	else
+		_confirm "$dest_file"
+		if [ "$confirmed" == "yes" ]; then
+			[ -e "$dest_file" -o -L "$dest_file" ] && rm "$dest_file"
+			ln -s "$src_file" "$dest_file"
+			echo "Installed $dest_file"
+		fi
 	fi
 }
